requirements.txt
Flask==3.1.2
peewee==3.18.2
openai==2.5.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn
google-cloud-translate==3.15.3
